Use single page for UCSF subcategory and eligibility. Move subcategory and eligibility pages into components and share code where appropriate.
