+++ app/components/ui/Navigation.module.scss
@@ -62,6 +62,7 @@
   @include navLogoShared;
 }
 
+.navLogoUcsf,
 .navLogoSFFamilies {
   margin-right: calc-em(45px);
   img {
