+++ app/pages/UcsfDiscoveryForm/UcsfDiscoveryForm.module.scss
@@ -1,0 +1,75 @@
+@import '~styles/utils/_helpers.scss';
+
+@import '~styles/utils/_helpers.scss';
+
+.eligibilityPage {
+  grid-column: 2 / span 10;
+
+  @media screen and (max-width: $break-tablet-s) {
+    grid-column: 1 / span 12;
+  }
+}
+
+.eligibilitiesBox {
+  display: flex;
+  margin: 16px 0;
+  padding: 32px;
+  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1), 0px 0px 4px rgba(0, 0, 0, 0.1);
+
+  @media screen and (max-width: $break-tablet-s) {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  &_title {
+    color: $color-brand;
+    @extend %font-size-large;
+    width: 86px;
+    margin-right: 26px;
+
+    @media screen and (max-width: $break-tablet-s) {
+      width: auto;
+      margin: auto auto 15px;
+    }
+  }
+}
+
+.eligibilitiesList {
+  display: grid;
+  gap: 8px;
+}
+
+.eligibilitiesLabels {
+  display: grid;
+  gap: 45px;
+  list-style: decimal inside;
+}
+
+.eligibilityGroupLabel {
+  display: list-item;
+}
+
+.eligibilityLabel {
+  cursor: pointer;
+  padding-left: 8px;
+}
+
+.listContainer {
+  display: grid;
+  grid-template-columns: 1fr 3fr;
+  gap: 30px;
+
+  @media screen and (max-width: $break-tablet-p) {
+    grid-template-columns: none;
+  }
+}
+
+.eligibilitiesBtns {
+  display: flex;
+  justify-content: right;
+  column-gap: 8px;
+
+  @media screen and (max-width: $break-tablet-s) {
+    justify-content: center;
+  }
+}
