+++ app/pages/UcsfServiceTypePage/UcsfServiceTypePage.tsx
@@ -1,17 +1,8 @@
-import React, { useState } from 'react';
-import { useHistory } from 'react-router-dom';
-import qs from 'qs';
-import ReactGA from 'react-ga';
+import React from 'react';
 
 import { Checkbox } from 'components/ui/inline/Checkbox/Checkbox';
-import { Button } from 'components/ui/inline/Button/Button';
-import { Section } from 'components/ucsf/Section/Section';
-import { Layout } from 'components/ucsf/Layout/Layout';
 
-import { constants } from './constants';
 import styles from './UcsfServiceTypePage.module.scss';
-
-import { useSubcategoriesForCategory } from '../../hooks/APIHooks';
 
 interface SubcategoryRefinement {
   name: string;
@@ -24,11 +15,16 @@ interface SelectedSubcategories {
 
 const seeAllPseudoId = -1;
 
-const ServiceTypes = ({ subcategories, selectedSubcategories, setSelectedSubcategories }: {
+export const ServiceTypes = ({ subcategories, selectedSubcategories, setSelectedSubcategories }: {
   subcategories: SubcategoryRefinement[];
   selectedSubcategories: SelectedSubcategories;
   setSelectedSubcategories: (categories: SelectedSubcategories) => void;
 }) => {
+  // Add generic "See All" element to subcategory array if it is not there yet
+  if (!subcategories[0] || subcategories[0].id !== seeAllPseudoId) {
+    subcategories.unshift({ id: seeAllPseudoId, name: 'See All' });
+  }
+
   const handleSubcategoryClick = (targetSubcategoryId: number) => {
     const seeAllIsTargetElement = targetSubcategoryId === seeAllPseudoId;
     const newValue = !selectedSubcategories[targetSubcategoryId];
@@ -73,96 +69,3 @@ const ServiceTypes = ({ subcategories, selectedSubcategories, setSelectedSubcate
     </div>
   );
 };
-
-const Page = () => {
-  const [selectedSubcategories, setSelectedSubcategories] = useState<SelectedSubcategories>({});
-  interface LocationState {
-    selectedResourceSlug: string;
-    selectedEligibilityNames: string[];
-  }
-
-  const history = useHistory<LocationState>();
-  const { state } = history.location;
-  const selectedResourceSlug = state && state.selectedResourceSlug;
-  const selectedEligibilities = state && state.selectedEligibilityNames;
-
-  const goToResourceResults = (slug: string) => {
-    const searchState = {
-      refinementList: {
-        eligibilities: selectedEligibilities,
-        categories: subcategories.flatMap(c => {
-          const isSeeAllItem = c.id === seeAllPseudoId;
-          if (!isSeeAllItem && selectedSubcategories[c.id]) {
-            return c.name;
-          }
-          return [];
-        }),
-      },
-    };
-
-    const categoriesRefinements = searchState.refinementList.categories.join('; ') || 'NONE';
-    const eligibilitiesRefinements = searchState.refinementList.eligibilities.join('; ') || 'NONE';
-    const search = qs.stringify(searchState, { encodeValuesOnly: true });
-
-    ReactGA.event({
-      category: 'UCSF Resource Inquiry',
-      action: 'Refined UCSF Resource Inquiry',
-      label: `${slug} Inquiry | Category Refinements: ${categoriesRefinements} | Eligibility Refinements: ${eligibilitiesRefinements}`,
-    });
-
-    history.push(`/${slug}/results?${search}`);
-  };
-
-  const backToEligibilityPage = (slug: string) => {
-    history.push('/client-identity', { selectedResourceSlug: slug, selectedEligibilityNames: [] });
-  };
-
-  if (!selectedResourceSlug) {
-    history.push('/');
-    return null;
-  }
-
-  const subcategories: SubcategoryRefinement[] = useSubcategoriesForCategory(
-    constants[selectedResourceSlug].id,
-  ) || [];
-
-  // Add generic "See All" element to subcategory array if it is not there yet
-  if (!subcategories[0] || subcategories[0].id !== seeAllPseudoId) {
-    subcategories.unshift({ id: seeAllPseudoId, name: 'See All' });
-  }
-
-  return (
-    <div className={styles.serviceTypePage}>
-      <Section
-        addClass={styles.subtitleMargin}
-        subtitle="Step 3: Can you tell us more about the services that your client is looking for?"
-      />
-      <div className={styles.serviceTypeContainer}>
-        <ServiceTypes
-          subcategories={subcategories}
-          selectedSubcategories={selectedSubcategories}
-          setSelectedSubcategories={setSelectedSubcategories}
-        />
-        <div className={styles.serviceTypeBtns}>
-          <Button
-            onClick={() => { backToEligibilityPage(selectedResourceSlug); }}
-          >
-            Back
-          </Button>
-          <Button
-            onClick={() => { goToResourceResults(selectedResourceSlug); }}
-            addClass={styles.goToResultsBtn}
-          >
-            Next: Service Capacity
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export const UcsfServiceTypePage = () => (
-  <Layout>
-    <Page />
-  </Layout>
-);
